requirements.txt
Django==1.7.6
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
PyYAML==3.11
Pygments==2.0.2
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
django-extensions==1.5.1
django-shell-plus==1.1.5
djangorestframework==3.1.0
funcy==1.4
gnureadline==6.3.3
ipython==3.0.0
jsonschema==2.4.0
mistune==0.5.1
numpy==1.9.2
pandas==0.15.2
ptyprocess==0.4
python-dateutil==2.4.1
pytz==2014.10
pyzmq==14.5.0
six==1.9.0
terminado==0.5
tornado==4.1
wsgiref==0.1.2
xlrd==0.9.3
